+++ Reach-Coder/Changes.md
@@ -4,6 +4,8 @@
 ### Bugfixes
 - Fixed bugged context property parsing
   - Previously the parsing would not recognize the static and variable variants
+### Other changes
+- Output is now compatible with **Reach v1.6**
 
 ## v1.0.3 - 23.1.2025
 This update adds support for the new context classes introduced in **Firmament v1.4**.